feat(*): Finished splitting classes, fixed test imports accordingly
